Changed some file names
