+++ scholarboard/src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
+import Navbar from './Components/Navbar';
 import Header from './Components/Header';
-import HomeTop from './Components/HomeTop';
 import Footer from './Components/Footer';
 import './Styles/styles.css';
 
@@ -9,8 +9,8 @@ class App extends Component {
   render() {
     return (
       <div className="App">
+        <Navbar />
         <Header />
-        <HomeTop />
         <Footer />
       </div>
     );
